requirements.txt
Django==1.6
django-bootstrap3==5.2.0
django-configurations==0.8
django-extensions==1.5.2
django-fab==1.0.4
django-model-utils==2.2
django-notifications-hq==0.6.2
ecdsa==0.13
fab-django-deploy==0.0.2
Fabric==1.10.1
paramiko==1.15.2
pip==6.1.1
psycopg2==2.6
pycrypto==2.6.1
selenium==2.45.0
setuptools==15.0
six==1.9.0
South==1.0.2
Werkzeug==0.10.4
